+++ packages/client/api/tsconfig.json
@@ -1,8 +1,7 @@
 {
     "extends": "../../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist",
-        "moduleResolution": "node"
+        "outDir": "./dist"
     },
     "exclude": ["node_modules", "dist"]
 }
