+++ packages/core/interface/package.json
@@ -1,5 +1,6 @@
 {
     "name": "@fluencelabs/interface",
+    "type": "module",
     "version": "0.1.0",
     "description": "Interfaces",
     "main": "./dist/index.js",
@@ -8,13 +9,30 @@
         ".": {
             "import": "./dist/index.js",
             "types": "./dist/index.d.ts"
+        },
+        "./fluenceClient": {
+            "import": "./dist/fluenceClient.js",
+            "types": "./dist/fluenceClient.d.ts"
+        },
+        "./compilerSupport": {
+            "import": "./dist/compilerSupport.js",
+            "types": "./dist/compilerSupport.d.ts"
+        }
+    },
+    "typesVersions": {
+        "*": {
+            "fluenceClient.d.ts": [
+                "./dist/fluenceClient.d.ts"
+            ],
+            "compilerSupport.d.ts": [
+                "./dist/compilerSupport.d.ts"
+            ]
         }
     },
     "engines": {
         "node": ">=10",
         "pnpm": ">=3"
     },
-    "type": "module",
     "scripts": {
         "build": "tsc"
     },
